Struct kana::Kana

pub struct Kana {
    // some fields omitted
}

Methods

impl Kana

fn init() -> Kana

impl Kana

fn half2full(&self, s: &str) -> String

Convert Half-width-kana into normal Katakana with diacritical marks separated [ｱﾞﾊﾟ -> ア゙パ]

This method is simple, but tends to cause troubles when rendering. In such a case, use half2kana() or execute vsmark2{half|full|combi}() as a post process.

Examples

use kana::Kana;
let k = Kana::init();
let s1 = "ﾏﾂｵ ﾊﾞｼｮｳ ｱﾟ";
assert_eq!("マツオ バショウ ア゚", k.half2full(s1));

fn half2kana(&self, s: &str) -> String

Convert Half-width-kana into normal Katakana with diacritical marks combined [ｱﾞﾊﾟ -> アﾞパ]

Examples

use kana::Kana;
let k = Kana::init();
let s1 = "ﾏﾂｵ ﾊﾞｼｮｳ ｱﾟ";
assert_eq!("マツオ バショウ ア ゚", k.half2kana(s1));

fn combine(&self, s: &str) -> String

Combine base characters and diacritical marks on Hiragana/Katakana [かﾞハ゜ -> がパ]

Examples

use kana::Kana;
let k = Kana::init();
let s2 = "ひ゜ひ゛んは゛";
assert_eq!("ぴびんば", k.combine(s2));